Experimental support for legacy / malformed responses from proxied server.
This is entirely to support behavior of broken web servers that might be
proxied.

The culprit, in this case, is a Jenkins server on Ubuntu 12.04 (installed using
their repositories and running in the default Servlet container). This server
for whatever reason generates responses that have this rare ill-formed set of
headers:

  HTTP/1.1 200 OK
  Connection: Close

With *no content-length* and *no transfer-encoding*.

So this is an initial attempt at working around this.