Support require-secure with remote SSL terminators
Previously require-secure would cause a redirect loop if run on an
app server behind an SSL-terminating reverse proxy, like an Amazon
ELB. That's because it ignored X-FORWARDED-PROTO and other headers
and just checked whether keter was being talked to over an encrypted
socket. This change uses WAI's "was this request made over a secure
connection?" logic as a supplement to the old way.