Force usage of http-reverse-proxy above 0.6.0.1
This prevents a DoS attack by using a head followed by
a post.
There is a test to confirm this on keter as well.

This includes several internal changes as well,
such as
Getting rid of HMState alias,
and
remove perform action & adjustbound from withclient where block

chop up the withclient function to make it easier to understand,
alsom makes the flow of bound values more explicit in type
signatures.