Fix the rest of the linting errors
Now all hlint and stylish-haskell have been fixed, and all flake checks
are passing.

One of the linters was failing to parse some of the files because of the
use of CPP for conditional imports. This is to maintain backwards
compatibility — which is of course a good thing — but I think it's the
wrong thing to do in this case.

1. I don't think it's worth maintaining backwards compatibility with
   http-reverse-proxy 0.4.2 which was released 10 years ago.

2. It's definitely not worth maintaining backwards compatibility with
   aeson < 2.0.1.0 since it fixed a hash collision vulnerability.